Fixup some tests
